// frontend/src/backend.d.ts
export enum Kategori {
    balita = "balita",
    tidakDitentukan = "tidakDitentukan",
    ibuMenyusui = "ibuMenyusui",
    ibuHamil = "ibuHamil"
}
export enum Status {
    aktif = "aktif",
    nonAktif = "nonAktif"
}